Model Wired Up to VR
It's really messy, right now, but I think it's a good idea to get this in and then clean it up in little bits.

Particularly, the mirrors can be merged, and the settings of the IK targets can be modified.

It may be worth moving this into a different scene since it's mainly for testing purposes, and much of what needs to be done now is small tweaks.